1, Comment internal visibility of test project 2, Add unit tests for VFloat

+++ Synchronica/Simulation/GameObject.cs
@@ -90,7 +90,7 @@ namespace Synchronica.Simulation
             return variable;
         }
 
-        public VFloat CreateFloat(short value)
+        public VFloat CreateFloat(float value)
         {
             var variable = new VFloat(GetNextVariableId(), value);
             this.variables.Add(variable);
